+++ src/plone/api/content.py
@@ -485,6 +485,32 @@ def transition(obj=None, transition=None, to_state=None, **kwargs):
             )
 
 
+@required_parameters('obj')
+def disable_acquire_local_roles(obj=None):
+    """Disable acquisition of local roles on given obj.
+    Set __ac_local_roles_block__ = 1 on obj.
+
+    :param obj: [required] Context object to block the acquisition on.
+    :type obj: Content object
+    :Example: :ref:`content_disable_acquire_local_roles_example`
+    """
+    plone_utils = portal.get_tool('plone_utils')
+    plone_utils.acquireLocalRoles(obj, status=0)
+
+
+@required_parameters('obj')
+def enable_acquire_local_roles(obj=None):
+    """Enable acquisition of local roles on given obj.
+    Set __ac_local_roles_block__ = 0 on obj.
+
+    :param obj: [required] Context object to enable the acquisition on.
+    :type obj: Content object
+    :Example: :ref:`content_enable_acquire_local_roles_example`
+    """
+    plone_utils = portal.get_tool('plone_utils')
+    plone_utils.acquireLocalRoles(obj, status=1)
+
+
 @required_parameters('name', 'context', 'request')
 def get_view(name=None, context=None, request=None):
     """Get a BrowserView object.